css files moved, some other changes
