+++ game2048/src/components/HelloComponent.js
@@ -1,13 +1,11 @@
-import "./HelloComponent.css";
+import "../styles/HelloComponent.css";
 
 const Hello = () => {
-    return (
-        <div className="App">
-      <header className="App-header">
-        Hallo
-      </header>
+  return (
+    <div className="App">
+      <header className="App-header">Hallo</header>
     </div>
-    );
+  );
 };
 
 export default Hello;
