+++ game2048/src/styles/HelloComponent.css
@@ -1,14 +1,14 @@
-.App {
+ .App {
     text-align: center;
   }
   
   .App-header {
     background-color: #282c34;
-    min-height: 100vh;
+    /* min-height: 100vh; */
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    font-size: calc(10px + 2vmin);
+    /* font-size: calc(10px + 2vmin); */
     color: white;
-  }
+  }
