+++ game2048/src/styles/Navbar.css
